refactor: extract archive notice to separate file
